# Use `AHashMap` for `classes` in `ClassDB`