`classes` is a major bottleneck for `ClassDB`, and much `Object` related functionality. A "proper" type system could correct this problem (https://github.com/godotengine/godot-proposals/issues/12323).

Eventually, `ClassDB` will not own class types (instead, `Object` subtypes will own them statically, and only register them in the `ClassDB`). To work towards this, this PR changes ownership from `HashMap` managed ownership to explicit allocations and deallocations.

As a side effect, lookups receive slightly improved efficiency (see benchmark).

## Benchmark

I used the same property benchmark as in https://github.com/godotengine/godot/pull/106646. While not nearly effective here as in that PR, we still see a ~3% improvement from GDScript. Considering this is a much less intrusive PR, the change is worth it for the performance improvement alone.

```c++
extends Camera3D

func _ready() -> void:
	var start := Time.get_ticks_usec()
	for i in range(5000000):
		editor_description = editor_description
		editor_description = editor_description
		editor_description = editor_description
		editor_description = editor_description
	var end := Time.get_ticks_usec()
	print((end - start) / 1000, "ms")
```

I ran the test 3 times (due to slight fluctuations):

```
before: 1824ms + 1799ms + 1811ms = 5.434
after: 1856ms + 1862ms + 1871ms = 5.589

```