@clayjohn 2025-10-01
Is your before and after reversed? Right now the numbers show a 3% regression in performance.