@DeeJayLSP 2025-10-29
Tested this on top of #112129 and #106646 using [Gaijin's Bunnymark](https://github.com/GaijinEntertainment/godot-das/tree/master/examples/bunnymark) at 500 bunnies.

FPS-wise, the results were quite mixed:

| Build     | Min FPS | Max FPS |
|-----------|---------|---------|
| Two above | 2054    | 2090    |
| Two + PR  | 2054    | 2087    |